FullLauncher: simplify the Drop code, and make it more likely to accept drops from other drag objects.

+++ qml/LaunchBar/DraggableAppIconDelegateModel.qml
@@ -40,8 +40,6 @@ VisualDataModel {
             property string modelId: model.id
             property string modelParams:  model.params === undefined ? "{}" : model.params
 
-            property variant placeHolderItem;
-
             LaunchableAppIcon {
                 id: launcherIcon
 
@@ -51,7 +49,7 @@ VisualDataModel {
                     radius: 5
                     border.color: "lightblue"
                     border.width: 2
-                    visible: model.title === ""
+                    visible: !model.title || model.title === ""
                 }
 
                 anchors {
@@ -99,10 +97,6 @@ VisualDataModel {
                 propagateComposedEvents: true
                 onPressAndHold: {
                     console.log("=== drag ===");
-
-                    // create a placeHolder for this delegate group
-                    createPlaceHolderAt(appsVisualDataModel, launcherIconDelegate.VisualDataModel.itemsIndex);
-
                     // move our delegate to the persisted items group
                     launcherIconDelegate.VisualDataModel.groups = [ "persistedItems" ];
 
@@ -111,93 +105,23 @@ VisualDataModel {
                 onReleased: {
                     if( held ) {
                         console.log("trigger drop");
-                        if( launcherIcon.Drag.target ) {
+                        if( launcherIcon.Drag.target && launcherIcon.Drag.target.placeHolderItem ) {
                             launcherIcon.Drag.drop();
                         }
                         else {
-                            // we are not on top of a drop target: fallback.
-                            dropAppIcon(launcherIconDelegate);
+                            console.log("no drop target, resetting drag source");
+                            launcherIconDelegate.VisualDataModel.itemsIndex = -1;
+                            launcherIconDelegate.VisualDataModel.inItems = true;
                         }
+
                         held = false;
 
-                        if( launcherIconDelegate.placeHolderItem ) {
-                            launcherIconDelegate.placeHolderItem.inItems = false;
-                            launcherIconDelegate.placeHolderItem = undefined;
-                        }
+                        // save that layout in DB
+                        launcherIconDelegate.VisualDataModel.model.saveCurrentLayout();
+
+                        launcherIconDelegate.VisualDataModel.inPersistedItems = false;
                     }
                 }
-            }
-
-            DropArea {
-                anchors { fill: parent; margins: 10 }
-
-                onEntered: {
-                    // Do nothing if dragging over the placeHolder item
-                    if( !launcherIconDelegate.VisualDataModel.isUnresolved && drag.source !=  launcherIconDelegate  ) {
-                        console.log("moving from " + drag.source.placeHolderItem.itemsIndex + " to " + launcherIconDelegate.VisualDataModel.itemsIndex);
-                        appsVisualDataModel.items.move(
-                                drag.source.placeHolderItem.itemsIndex,
-                                launcherIconDelegate.VisualDataModel.itemsIndex);
-                    }
-                }
-                onDropped: {
-                    dropAppIcon(drag.source);
-                }
-            }
-
-            function createPlaceHolderAt(targetAppsVisualDataModel, atIndex)
-            {
-                if( launcherIconDelegate.placeHolderItem ) {
-                    console.log("removing old placeHolder");
-                    launcherIconDelegate.placeHolderItem.inItems = false;
-                    launcherIconDelegate.placeHolderItem = undefined;
-                }
-                targetAppsVisualDataModel.items.insert(atIndex, {title: "", icon: "", id: ""});
-                launcherIconDelegate.placeHolderItem = targetAppsVisualDataModel.items.get(atIndex);
-                launcherIconDelegate.placeHolderItem.sourceProperties = {
-                     title: launcherIconDelegate.modelTitle,
-                     icon: launcherIconDelegate.modelIcon,
-                     id: launcherIconDelegate.modelId,
-                     params: launcherIconDelegate.modelParams
-                };
-                launcherIconDelegate.placeHolderItem.appsVisualDataModel = targetAppsVisualDataModel;
-
-                console.log("placeHolderItemIndex = " + atIndex + ", launcherIconDelegate.placeHolderItem = " + launcherIconDelegate.placeHolderItem);
-            }
-
-            function dropAppIcon(sourceDrag)
-            {
-                console.log("=== drop === ");
-
-                // replace the placeHolder with ourself
-                var placeHolderItemsIndex = sourceDrag.placeHolderItem.itemsIndex;
-                var placeHolderVisualDataModel = sourceDrag.placeHolderItem.appsVisualDataModel;
-
-                console.log("placeHolderItemsIndex ="+placeHolderItemsIndex);
-                if( sourceDrag.VisualDataModel.model === placeHolderVisualDataModel ) {
-                    // same tab
-                    console.log("move dragged icon from = " + sourceDrag.VisualDataModel.itemsIndex);
-                    sourceDrag.VisualDataModel.inItems = true;
-                    placeHolderVisualDataModel.items.move(sourceDrag.VisualDataModel.itemsIndex, placeHolderItemsIndex);
-                    console.log("moved dragged icon to = " + sourceDrag.VisualDataModel.itemsIndex);
-
-                    // save that layout in DB (it will save the whole launcher)
-                    placeHolderVisualDataModel.saveCurrentLayout();
-                }
-                else {
-                    // different tab
-                    placeHolderVisualDataModel.items.insert(placeHolderItemsIndex, {
-                                                         title: sourceDrag.placeHolderItem.sourceProperties.title,
-                                                         id: sourceDrag.placeHolderItem.sourceProperties.id,
-                                                         icon: sourceDrag.placeHolderItem.sourceProperties.icon,
-                                                         params: sourceDrag.placeHolderItem.sourceProperties.params });
-
-                    // save that layout in DB (it will save the whole launcher)
-                    placeHolderVisualDataModel.saveCurrentLayout();
-                    sourceDrag.VisualDataModel.model.saveCurrentLayout();
-                }
-
-                sourceDrag.VisualDataModel.inPersistedItems = false;
             }
         }
 }
